+++ R/geoflow_relation.R
@@ -14,6 +14,7 @@
 #' @examples
 #' \dontrun{
 #'   relation <- geoflow_relation$new()
+#'   relation$setProv("user")
 #'   relation$setKey("wms")
 #'   relation$setLink("http://somelink/wms")
 #'   relation$setMimeType("application/xml")
@@ -35,6 +36,8 @@ geoflow_relation <- R6Class("geoflow_relation",
    name = NULL,
    #'@field description relation name
    description = NULL,
+   #'@field prov provenance,
+   prov = "user",
    
    #'@description Initializes an object of class \link{geoflow_relation}
    #'@param str character string to initialize from using key-based syntax
@@ -90,6 +93,13 @@ geoflow_relation <- R6Class("geoflow_relation",
    #'@param description description
    setDescription = function(description){
      self$description <- description
+   },
+   
+   #'@description Set provenance
+   #'@param prov provenance
+   setProv = function(prov = c("user","geoflow")){
+     prov = match.arg(prov)
+     self$prov = prov
    }
  )                                  
 )